refactor(#43): refactors the code base
