+++ GalacticGuardians/android/src/com/tdt4240gr18/game/AndroidLauncher.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
 import com.google.firebase.FirebaseApp;
+import com.tdt4240gr18.GalacticGuardians;
 
 public class AndroidLauncher extends AndroidApplication {
 	private final AndroidDatabaseHandler databaseHandler = new AndroidDatabaseHandler();
