+++ GalacticGuardians/core/src/com/tdt4240gr18/game/misc/Option.java
@@ -1,4 +1,4 @@
-package com.tdt4240gr18.game;
+package com.tdt4240gr18.game.misc;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
